refactor : move package ManualHandlerAdapter
